+++ src/redux/modules/userData/selectors.js
@@ -3,17 +3,18 @@
  */
 
 import { createSelector } from 'reselect';
+import { hasAllBranding } from '../entities/branding/selectors';
 
 const getUserData = state => state.get('UserData');
 
-export const getCurrentTenantName = state =>
-  state.get('UserData').get('currentTenantName');
+export const getCurrentTenantName = state => state.get('UserData').get('currentTenantName');
+
+export const userIsAuthed = state => state.get('UserData').get('userIsAuthed');
 
 export const getCurrentAgentId = state => state.get('UserData').get('agentId');
 
-export const getCurrentTenantId = state =>
-  state.get('UserData').get('currentTenantId');
+export const getCurrentTenantId = state => state.get('UserData').get('currentTenantId');
 
-export const getCurrentPermissions = createSelector([getUserData], userData =>
-  userData.get('permissions')
-);
+export const getCurrentPermissions = createSelector([getUserData], userData => userData.get('permissions'));
+
+export const authenticatedAndBrandingReady = state => hasAllBranding(state) && userIsAuthed(state);